refactor: organize CategoriesList component directory
