+++ components/CategoriesList/CategoryCard.styles.ts
